chore(main): move injectExtensions to util function
