+++ src/main/util/installExtensions.js
@@ -1,0 +1,13 @@
+/* eslint-disable global-require, import/no-extraneous-dependencies */
+
+export default function installExtensions() {
+  const {
+    default: installer,
+    REACT_DEVELOPER_TOOLS,
+    REDUX_DEVTOOLS
+  } = require('electron-devtools-installer');
+
+  const extensions = [REACT_DEVELOPER_TOOLS, REDUX_DEVTOOLS];
+
+  return Promise.all(extensions.map((extension) => installer(extension)));
+}
